+++ src/biome/text/api_new/helpers.py
@@ -12,7 +12,7 @@ from allennlp.modules.seq2seq_encoders import PytorchSeq2SeqWrapper, _Seq2SeqWra
 from allennlp.modules.seq2vec_encoders import PytorchSeq2VecWrapper, _Seq2VecWrapper
 from elasticsearch import Elasticsearch
 
-from . import constants, environment
+from biome.text.api_new import environment
 
 
 def yaml_to_dict(filepath: str) -> Dict[str, Any]:
@@ -51,69 +51,6 @@ def get_env_cuda_device() -> int:
     """
     cuda_device = int(os.getenv(environment.CUDA_DEVICE, "-1"))
     return cuda_device
-
-
-class ElasticsearchExplore:
-    """Elasticsearch data exploration class"""
-
-    def __init__(self, es_index: str, es_host: Optional[str] = None):
-        self.es_index = es_index
-        self.es_host = es_host or constants.DEFAULT_ES_HOST
-        if not self.es_host.startswith("http"):
-            self.es_host = f"http://{self.es_host}"
-
-        self.client = Elasticsearch(
-            hosts=es_host, retry_on_timeout=True, http_compress=True
-        )
-        self.es_doc = get_compatible_doc_type(self.client)
-
-    def create_explore_data_record(self, parameters: Dict[str, Any]):
-        """Creates an exploration data record data exploration"""
-
-        self.client.indices.create(
-            index=constants.BIOME_METADATA_INDEX,
-            body={
-                "settings": {"index": {"number_of_shards": 1, "number_of_replicas": 0}}
-            },
-            params=dict(ignore=400),
-        )
-
-        self.client.update(
-            index=constants.BIOME_METADATA_INDEX,
-            doc_type=constants.BIOME_METADATA_INDEX_DOC,
-            id=self.es_index,
-            body={
-                "doc": dict(
-                    name=self.es_index, created_at=datetime.datetime.now(), **parameters
-                ),
-                "doc_as_upsert": True,
-            },
-        )
-
-    def create_explore_data_index(self, force_delete: bool):
-        """Creates an explore data index if not exists or is forced"""
-        dynamic_templates = [
-            {
-                data_type: {
-                    "match_mapping_type": data_type,
-                    "path_match": path_match,
-                    "mapping": {
-                        "type": "text",
-                        "fields": {"keyword": {"type": "keyword", "ignore_above": 256}},
-                    },
-                }
-            }
-            for data_type, path_match in [("*", "*.value"), ("string", "*")]
-        ]
-
-        if force_delete:
-            self.client.indices.delete(index=self.es_index, ignore=[400, 404])
-
-        self.client.indices.create(
-            index=self.es_index,
-            body={"mappings": {self.es_doc: {"dynamic_templates": dynamic_templates}}},
-            ignore=400,
-        )
 
 
 def method_argument_types(the_method: Callable) -> Dict[str, Type[Any]]:
