+++ src/biome/text/api_new/modules/encoders/__init__.py
@@ -1,4 +1,6 @@
 from biome.text.api_new.modules.specs import Seq2SeqEncoderSpec
-from .time_distributed_encoder import TimeDistributedEncoder
+from biome.text.api_new.modules.encoders.time_distributed_encoder import (
+    TimeDistributedEncoder,
+)
 
 Encoder = Seq2SeqEncoderSpec
